requirements.txt
torch
datasets 
transformers
peft 
bitsandbytes
pyyaml
matplotlib
pandas